ai-diagnostic-assistant/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
anthropic==0.34.2
pydantic==2.9.2
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
httpx==0.27.2
